requirements.txt
fastapi
uvicorn
tortoise
tortoise-orm
aerich
pydantic
typing
statics
starlette
asyncpg
fastapi-pagination
PyJWT